+++ src/components/CommunityMain/SectionExhibition.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from 'next/image';
 import MoveBtn from './MoveBtn';
-import * as S from './main-styled';
+import * as S from './main.styles';
 import { EXHIBITION_DATE } from '../../constants/communityMain';
 
 const SectionExhibition = () => {
