+++ src/components/CommunityMain/SectionStory.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from 'next/image';
-import * as S from './main-styled';
+import * as S from './main.styles';
 import { STORY_DATA, STORY_LIST } from '../../constants/communityMain';
 
 const SectionStory = () => {
