+++ src/components/CommunityMain/main.styles.js
